extracted drawer widget from home page
